requirements.txt
numpy
pandas
scikit-learn
scipy
joblib
seaborn
matplotlib
statsmodels
ipykernel
transformers
torch
textstat
accelerate
python-dotenv
dataclasses_json
tables
loguru
langchain-community
langchain-ollama
tolerantjson
